WIP cleanup box event handler
