+++ box-service/Cargo.toml
@@ -29,7 +29,7 @@ jsonwebtoken = "9.2.0"
 base64 = "0.21.7"
 axum-auth = "0.4.0"
 # Include the shared crate
-lockbox-shared = { path = "../shared" }
+lockbox-shared = { path = "../shared", features = ["test_utils"] }
 
 [dev-dependencies]
 mockito = "1.3.0" # For mocking HTTP requests/responses if needed later
