+++ box-service/src/main.rs
@@ -1,8 +1,9 @@
 mod error;
 mod handlers;
+// Keep models for request/response types
 mod models;
 mod routes;
-mod store;
+
 #[cfg(test)]
 mod tests;
 
@@ -12,6 +13,10 @@ use lambda_http::{
     Response as LambdaResponse,
 };
 use tower::ServiceExt;
+
+// Export shared models and store for the rest of the crate to use
+pub use lockbox_shared::models as shared_models;
+pub use lockbox_shared::store;
 
 // The Lambda handler function
 async fn function_handler(event: LambdaRequest) -> Result<LambdaResponse<LambdaBody>, Error> {
