+++ box-service/src/tests/mod.rs
@@ -1,3 +1,2 @@
 pub mod box_tests;
 pub mod guardian_tests;
-pub mod store_tests;
