+++ invitation-service/Cargo.toml
@@ -22,6 +22,7 @@ thiserror = "1.0"
 lambda_http = "0.9.0"
 aws-config = "1.3.1"
 aws-sdk-dynamodb = ">=1.0.0"
+aws-sdk-sns = "1.3.1"
 serde_dynamo = { version = "4.2.0", features = ["aws-sdk-dynamodb+1"] }
 async-trait = "0.1.77"
 aws_lambda_events = { version = "0.11", default-features = false, features = ["apigw"] }
@@ -29,8 +30,9 @@ jsonwebtoken = "9.2.0"
 base64 = "0.21.7"
 axum-auth = "0.4.0"
 nanoid = "0.4.0"
+anyhow = "1.0"
 # Include the shared crate
-lockbox-shared = { path = "../shared" }
+lockbox-shared = { path = "../shared", features = ["test_utils"] }
 
 [dev-dependencies]
 mockito = "1.3.0" # For mocking HTTP requests/responses if needed later
