+++ shared/src/models/mod.rs
@@ -1,6 +1,7 @@
 use chrono::Utc;
 use serde::{Deserialize, Serialize};
 
+// Invitation-related models
 #[derive(Serialize, Deserialize, Clone, Debug)]
 pub struct Invitation {
     pub id: String,
@@ -20,6 +21,92 @@ pub struct Invitation {
     pub creator_id: String, // ID of the user who created the invitation
 }
 
+// Box-related models
+#[derive(Serialize, Deserialize, Clone, Debug)]
+pub struct Document {
+    pub id: String,
+    pub title: String,
+    pub content: String,
+    #[serde(rename = "createdAt")]
+    pub created_at: String,
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug)]
+pub struct Guardian {
+    pub id: String, // user_id
+    pub name: String,
+    pub lead: bool,
+    pub status: String, // "sent", "viewed", "accepted", "rejected"
+    #[serde(rename = "addedAt")]
+    pub added_at: String,
+    #[serde(rename = "invitationId")]
+    pub invitation_id: String,
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug)]
+pub struct UnlockRequest {
+    pub id: String,
+    #[serde(rename = "requestedAt")]
+    pub requested_at: String,
+    pub status: String,
+    pub message: Option<String>,
+    #[serde(rename = "initiatedBy")]
+    pub initiated_by: Option<String>,
+    #[serde(rename = "approvedBy")]
+    pub approved_by: Vec<String>,
+    #[serde(rename = "rejectedBy")]
+    pub rejected_by: Vec<String>,
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug)]
+pub struct BoxRecord {
+    pub id: String,
+    pub name: String,
+    pub description: String,
+    pub is_locked: bool,
+    pub created_at: String,
+    pub updated_at: String,
+    pub owner_id: String,
+    pub owner_name: Option<String>,
+    pub documents: Vec<Document>,
+    pub guardians: Vec<Guardian>,
+    pub lead_guardians: Vec<Guardian>,
+    pub unlock_instructions: Option<String>,
+    pub unlock_request: Option<UnlockRequest>,
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+pub struct GuardianBox {
+    pub id: String,
+    pub name: String,
+    pub description: String,
+    #[serde(rename = "isLocked")]
+    pub is_locked: bool,
+    #[serde(rename = "createdAt")]
+    pub created_at: String,
+    #[serde(rename = "updatedAt")]
+    pub updated_at: String,
+    #[serde(rename = "ownerId")]
+    pub owner_id: String,
+    #[serde(rename = "ownerName")]
+    pub owner_name: Option<String>,
+    #[serde(rename = "unlockInstructions")]
+    pub unlock_instructions: Option<String>,
+    #[serde(rename = "unlockRequest")]
+    pub unlock_request: Option<UnlockRequest>,
+    #[serde(rename = "pendingGuardianApproval")]
+    pub pending_guardian_approval: Option<bool>,
+    #[serde(rename = "guardiansCount")]
+    pub guardians_count: usize,
+    #[serde(rename = "isLeadGuardian")]
+    pub is_lead_guardian: bool,
+    // TODO we probably shouldn't be just returning them all for privacy reasons
+    pub documents: Vec<Document>,
+    pub guardians: Vec<Guardian>,
+    #[serde(rename = "leadGuardians")]
+    pub lead_guardians: Vec<Guardian>,
+}
+
 // Response DTOs for general use across services
 #[derive(Serialize, Debug)]
 pub struct ErrorResponse {
@@ -29,6 +116,8 @@ pub struct ErrorResponse {
 #[derive(Serialize, Debug)]
 pub struct MessageResponse {
     pub message: String,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub box_id: Option<String>,
 }
 
 // Helper function to get current timestamp as string
