+++ shared/src/store/dynamo.rs
@@ -1,5 +1,10 @@
 use async_trait::async_trait;
 use aws_config::BehaviorVersion;
+use aws_sdk_dynamodb::error::SdkError;
+use aws_sdk_dynamodb::operation::delete_item::DeleteItemError;
+use aws_sdk_dynamodb::operation::get_item::GetItemError;
+use aws_sdk_dynamodb::operation::query::QueryError;
+use aws_sdk_dynamodb::operation::scan::ScanError;
 use aws_sdk_dynamodb::types::AttributeValue;
 use aws_sdk_dynamodb::Client;
 use chrono::{Duration, Utc};
@@ -8,12 +13,19 @@ use std::collections::HashMap;
 use std::env;
 
 use crate::error::{map_dynamo_error, Result, StoreError};
-use crate::models::Invitation;
+use crate::models::{BoxRecord, Invitation, now_str};
 
+// Invitation Store Constants
 const TABLE_NAME: &str = "invitation-table";
 const GSI_BOX_ID: &str = "box_id-index";
 const GSI_INVITE_CODE: &str = "invite_code-index";
 const GSI_CREATOR_ID: &str = "creator_id-index";
+
+// Box Store Constants
+const BOX_TABLE_NAME: &str = "box-table";
+const GSI_OWNER_ID: &str = "owner_id-index";
+
+// DynamoInvitationStore
 
 pub struct DynamoInvitationStore {
     client: Client,
@@ -42,6 +54,188 @@ impl DynamoInvitationStore {
     }
 }
 
+// DynamoBoxStore
+
+/// DynamoDB store for boxes
+pub struct DynamoBoxStore {
+    client: Client,
+    table_name: String,
+}
+
+impl DynamoBoxStore {
+    /// Creates a new DynamoDB store
+    pub async fn new() -> Self {
+        // Use the recommended defaults() function with latest behavior version
+        let config = aws_config::defaults(BehaviorVersion::latest()).load().await;
+
+        let client = Client::new(&config);
+
+        // Use environment variable for table name if available
+        let table_name =
+            env::var("DYNAMODB_TABLE").unwrap_or_else(|_| BOX_TABLE_NAME.to_string());
+
+        Self { client, table_name }
+    }
+
+    /// Creates a new DynamoDB store with the specified client and table name.
+    /// This is mainly useful for testing with a local DynamoDB instance.
+    #[allow(dead_code)]
+    pub fn with_client_and_table(client: Client, table_name: String) -> Self {
+        Self { client, table_name }
+    }
+}
+
+#[async_trait]
+impl super::BoxStore for DynamoBoxStore {
+    /// Creates a new box record in DynamoDB
+    async fn create_box(&self, box_record: BoxRecord) -> Result<BoxRecord> {
+        let item = to_item(&box_record)?;
+
+        self.client
+            .put_item()
+            .table_name(&self.table_name)
+            .set_item(Some(item))
+            .send()
+            .await
+            .map_err(|e| map_dynamo_error("put_item", e))?;
+
+        Ok(box_record)
+    }
+
+    /// Gets a box by ID
+    async fn get_box(&self, id: &str) -> Result<BoxRecord> {
+        let key = HashMap::from([("id".to_string(), AttributeValue::S(id.to_string()))]);
+
+        let response = self
+            .client
+            .get_item()
+            .table_name(&self.table_name)
+            .set_key(Some(key))
+            .send()
+            .await
+            .map_err(|e| map_get_dynamo_error(e, id))?;
+
+        let item = response
+            .item()
+            .ok_or_else(|| StoreError::NotFound(format!("Box not found: {}", id)))?;
+
+        let box_record = from_item(item.clone())?;
+        Ok(box_record)
+    }
+
+    /// Gets all boxes owned by a user
+    async fn get_boxes_by_owner(&self, owner_id: &str) -> Result<Vec<BoxRecord>> {
+        let expr_attr_names = HashMap::from([("#owner_id".to_string(), "owner_id".to_string())]);
+
+        let expr_attr_values = HashMap::from([(
+            ":owner_id".to_string(),
+            AttributeValue::S(owner_id.to_string()),
+        )]);
+
+        let response = self
+            .client
+            .query()
+            .table_name(&self.table_name)
+            .index_name(GSI_OWNER_ID) // Use the GSI
+            .key_condition_expression("#owner_id = :owner_id")
+            .set_expression_attribute_names(Some(expr_attr_names))
+            .set_expression_attribute_values(Some(expr_attr_values))
+            .send()
+            .await
+            .map_err(|e| map_query_dynamo_error(e))?;
+
+        // items() returns a reference to a slice, which could be empty but not None
+        let items = response.items();
+
+        let mut boxes = Vec::new();
+        for item in items {
+            let box_record = from_item(item.clone())?;
+            boxes.push(box_record);
+        }
+
+        Ok(boxes)
+    }
+
+    /// Updates a box
+    async fn update_box(&self, box_record: BoxRecord) -> Result<BoxRecord> {
+        // For updates, simply use put_item to replace the entire item
+        // In a production app, you might want to use update_item with expressions for efficiency
+        let updated_box = BoxRecord {
+            updated_at: now_str(),
+            ..box_record
+        };
+
+        let item = to_item(&updated_box)?;
+
+        self.client
+            .put_item()
+            .table_name(&self.table_name)
+            .set_item(Some(item))
+            .send()
+            .await
+            .map_err(|e| map_dynamo_error("put_item", e))?;
+
+        Ok(updated_box)
+    }
+
+    /// Deletes a box
+    async fn delete_box(&self, id: &str) -> Result<()> {
+        let key = HashMap::from([("id".to_string(), AttributeValue::S(id.to_string()))]);
+
+        self.client
+            .delete_item()
+            .table_name(&self.table_name)
+            .set_key(Some(key))
+            .send()
+            .await
+            .map_err(|e| map_delete_dynamo_error(e))?;
+
+        Ok(())
+    }
+
+    /// Gets all boxes where the given user is a guardian (with status not rejected)
+    ///
+    /// Implementation notes:
+    /// - Currently uses a full table scan since guardians are stored in nested arrays within the BoxRecord
+    /// - For production systems with many boxes, this could be improved by:
+    ///   1. Creating a new GSI with a composite key or
+    ///   2. Creating a separate guardian-to-box mapping table with a GSI
+    ///   3. Using DynamoDB's new document path capabilities for filtering
+    async fn get_boxes_by_guardian_id(&self, guardian_id: &str) -> Result<Vec<BoxRecord>> {
+        // Currently we perform a full table scan as guardian information is stored in an array within
+        // the box document, not as a separate attribute that can be indexed. In the future, we could
+        // create a separate table or GSI for guardian relationships.
+
+        let response = self
+            .client
+            .scan()
+            .table_name(&self.table_name)
+            .send()
+            .await
+            .map_err(|e| map_scan_dynamo_error(e))?;
+
+        let items = response.items();
+
+        let mut boxes = Vec::new();
+        for item in items {
+            let box_record: BoxRecord = from_item(item.clone())?;
+
+            // Check if the user is a guardian for this box
+            let is_guardian = box_record
+                .guardians
+                .iter()
+                .any(|guardian| guardian.id == guardian_id && guardian.status != "rejected");
+
+            if is_guardian {
+                boxes.push(box_record);
+            }
+        }
+
+        Ok(boxes)
+    }
+}
+
+// INVITATION STORE IMPLEMENTATION
 #[async_trait]
 impl super::InvitationStore for DynamoInvitationStore {
     async fn create_invitation(&self, mut invitation: Invitation) -> Result<Invitation> {
@@ -210,9 +404,10 @@ impl super::InvitationStore for DynamoInvitationStore {
 
     async fn get_invitations_by_creator_id(&self, creator_id: &str) -> Result<Vec<Invitation>> {
         // Create expression attribute values
-        let expr_attr_values = HashMap::from([
-            (":creator_id".to_string(), AttributeValue::S(creator_id.to_string()))
-        ]);
+        let expr_attr_values = HashMap::from([(
+            ":creator_id".to_string(),
+            AttributeValue::S(creator_id.to_string()),
+        )]);
 
         let result = self
             .client
@@ -230,19 +425,37 @@ impl super::InvitationStore for DynamoInvitationStore {
         let mut invitations = Vec::new();
         for item in items {
             let invitation: Invitation = from_item(item.clone())?;
-            // Filter out expired invitations
-            let expires_at =
-                chrono::DateTime::parse_from_rfc3339(&invitation.expires_at).map_err(|_| {
-                    StoreError::InternalError("Invalid expiration date format".to_string())
-                })?;
-
-            if Utc::now() <= expires_at {
-                invitations.push(invitation);
-            }
+            invitations.push(invitation);
         }
 
         Ok(invitations)
     }
+}
+
+// Helper functions for DynamoDB error mapping
+fn map_get_dynamo_error(err: SdkError<GetItemError>, id: &str) -> StoreError {
+    match err {
+        SdkError::ServiceError(ref service_err) => {
+            if let GetItemError::ResourceNotFoundException(_) = service_err.err() {
+                StoreError::NotFound(format!("Box not found: {}", id))
+            } else {
+                StoreError::InternalError(format!("DynamoDB get_item error: {}", err))
+            }
+        }
+        _ => StoreError::InternalError(format!("DynamoDB get_item error: {}", err)),
+    }
+}
+
+fn map_delete_dynamo_error(err: SdkError<DeleteItemError>) -> StoreError {
+    StoreError::InternalError(format!("DynamoDB delete_item error: {}", err))
+}
+
+fn map_query_dynamo_error(err: SdkError<QueryError>) -> StoreError {
+    StoreError::InternalError(format!("DynamoDB query error: {}", err))
+}
+
+fn map_scan_dynamo_error(err: SdkError<ScanError>) -> StoreError {
+    StoreError::InternalError(format!("DynamoDB scan error: {}", err))
 }
 
 // Add Default impl for convenience
